Fixed printf warnings about data type not matching format expression.

+++ src/libxsmm_main.c
@@ -5225,7 +5225,7 @@ LIBXSMM_API_INTERN void libxsmm_matrix_eqn_trv_print( libxsmm_matrix_eqn_elem* c
   } else if ( cur_node->type == LIBXSMM_MATRIX_EQN_NODE_UNARY ) {
     /* we have to push more in this branch */
     if ( cur_node->le != NULL ) {
-      printf("UNARY: %i %i (timestamp = %i, tmp = %i)\n", cur_node->info.u_op.type, cur_node->info.u_op.flags, cur_node->visit_timestamp, cur_node->tmp_id );
+      printf("UNARY: %i %i (timestamp = %i, tmp = %i)\n", (int)cur_node->info.u_op.type, (int)cur_node->info.u_op.flags, cur_node->visit_timestamp, cur_node->tmp_id );
       libxsmm_matrix_eqn_trv_print( cur_node->le, indent+tree_print_indent );
     /* we have reached the root, as we are unary, there is no right branch */
     } else if ( (cur_node->ri != NULL) ) {
@@ -5234,7 +5234,7 @@ LIBXSMM_API_INTERN void libxsmm_matrix_eqn_trv_print( libxsmm_matrix_eqn_elem* c
   } else if ( cur_node->type == LIBXSMM_MATRIX_EQN_NODE_BINARY ) {
     /* we have to push more in this branch */
     if ( (cur_node->le != NULL) && (cur_node->ri != NULL) ) {
-      printf("BINARY: %i %i (timestamp = %i, tmp = %i)\n", cur_node->info.b_op.type, cur_node->info.b_op.flags, cur_node->visit_timestamp, cur_node->tmp_id );
+      printf("BINARY: %i %i (timestamp = %i, tmp = %i)\n", (int)cur_node->info.b_op.type, (int)cur_node->info.b_op.flags, cur_node->visit_timestamp, cur_node->tmp_id );
       libxsmm_matrix_eqn_trv_print( cur_node->le, indent+tree_print_indent );
       libxsmm_matrix_eqn_trv_print( cur_node->ri, indent+tree_print_indent );
     } else {
@@ -5259,7 +5259,7 @@ LIBXSMM_API_INTERN void libxsmm_matrix_eqn_trv_rpn_print( libxsmm_matrix_eqn_ele
     /* we have to push more in this branch */
     if ( cur_node->le != NULL ) {
       libxsmm_matrix_eqn_trv_rpn_print( cur_node->le );
-      printf("UNARY-%i ", cur_node->info.u_op.type );
+      printf("UNARY-%i ", (int)cur_node->info.u_op.type );
     /* we have reached the root, as we are unary, there is no right branch */
     } else if ( (cur_node->ri != NULL) ) {
       printf("ERROR: Unary cannot have right childs!\n");
@@ -5269,7 +5269,7 @@ LIBXSMM_API_INTERN void libxsmm_matrix_eqn_trv_rpn_print( libxsmm_matrix_eqn_ele
     if ( (cur_node->le != NULL) && (cur_node->ri != NULL) ) {
       libxsmm_matrix_eqn_trv_rpn_print( cur_node->le );
       libxsmm_matrix_eqn_trv_rpn_print( cur_node->ri );
-      printf("BINARY-%i ", cur_node->info.b_op.type );
+      printf("BINARY-%i ", (int)cur_node->info.b_op.type );
     } else {
       printf("ERROR: Binary needs left and right child!\n");
     }
